NuActionGUI/NuActionGUI/src/requirements.txt
# This file is used by pip to install required python packages
# Usage: pip install -r requirements.txt

pytest==7.2.0

# Flask Framework
Flask==2.2.2

# Flask Packages
Flask-Cors==3.0.4
Flask-SQLAlchemy==3.0.2
Flask-User==1.0.2.2
email_validator==1.3.0
Werkzeug==2.2.2

# Antlr parser
antlr4-python3-runtime==4.13.0
antlr4-tools==0.2

# CPython modifier
forbiddenfruit==0.1.4


